+++ reTurn/client/TurnAsyncTcpSocket.hxx
@@ -16,7 +16,7 @@ public:
                                const asio::ip::address& address = UnspecifiedIpAddress, 
                                unsigned short port = 0);
 
-   virtual unsigned int getSocketDescriptor() { return mSocket.native(); }
+   virtual unsigned int getSocketDescriptor() { return (unsigned int)mSocket.native(); }
 
 protected:
