+++ reTurn/client/TurnAsyncUdpSocket.hxx
@@ -16,7 +16,7 @@ public:
                                const asio::ip::address& address, 
                                unsigned short port); 
 
-   virtual unsigned int getSocketDescriptor() { return mSocket.native(); }
+   virtual unsigned int getSocketDescriptor() { return (unsigned int)mSocket.native(); }
 
 protected:
